// src/lib/touchline/types.ts
export interface Platform {
  id: number;
  name: string;
  enabled: boolean;
}

export interface ContentType {
  id: number;
  platform_id: number;
  name: string;
}

export interface SocialParameter {
  id: number;
  platform_id: number;
  content_type_id?: number;
  parameter_type: string;
  parameter_key: string;
  value: number;
}